Improve keepalive handling.

+++ api/src/main/java/net/elytrium/limboapi/api/player/LimboPlayer.java
@@ -71,5 +71,5 @@ public interface LimboPlayer {
 
   Player getProxyPlayer();
 
-  long getPing();
+  int getPing();
 }
